End-of-session bug-hunt + docs
Bug fixes:
- compute_full_metrics.py: same MCC int-overflow as fixed in 90bbecb
  (sibling dead-code copy, fixed defensively)
- consolidate_encoder_data.py: same fire_clim future-year leak pattern
  as fixed in train_v3.py — min(keys, key=abs(y-yr)) could pick yr+1
  → fix to max([y for y in keys if y <= yr])
- 6 SLURM scripts had 'echo "exit=\$?"' without explicit 'exit \$' —
  this masked Python crashes as COMPLETED (root cause of metric_card
  silent fail). Capture exit code into variable, then re-raise.

Docs:
- SOTA_ATTACKS_2026_05_10.md: full log of phase 5 (every experiment +
  outcome since 2026-04-25), with verdicts and what-not-to-retry list
- BUG_HUNT_2026_05_10.md: end-of-session sweep, 5 bug classes audited,
  3 fixed, 2 documented as tech debt
- RESULTS_TABLE_2026_05_03.md: §0 update with ensemble SOTA, 22y death,
  dm sweep, CWFIS drift @ 12y finding, NBAC 2025 gap

Tech debt left:
- 7 processing scripts still pass src_nodata=np.nan to reproject (edge
  hole risk per DATA_CONVENTIONS §7a) — not fixed because re-running
  would rebuild 309GB+ caches
- sub4x4 collate shape mismatch — not fixed, needs dataset code work

+++ scripts/compute_full_metrics.py
@@ -126,10 +126,13 @@ def mcc_at_k(score_flat, label_flat, k):
     order = np.argsort(score_flat)[::-1]
     pred = np.zeros(n, dtype=np.uint8)
     pred[order[:k]] = 1
-    tp = int(((pred == 1) & (label_flat == 1)).sum())
-    tn = int(((pred == 0) & (label_flat == 0)).sum())
-    fp = int(((pred == 1) & (label_flat == 0)).sum())
-    fn = int(((pred == 0) & (label_flat == 1)).sum())
+    # Cast to float64 — Python int multiplication overflows int64 here
+    # (tn ~ 6M for all-Canada → product ~ 1e26). np.sqrt also fails on
+    # Python int. Same fix as compute_full_metric_card.py / compute_unified_metrics.py.
+    tp = float(((pred == 1) & (label_flat == 1)).sum())
+    tn = float(((pred == 0) & (label_flat == 0)).sum())
+    fp = float(((pred == 1) & (label_flat == 0)).sum())
+    fn = float(((pred == 0) & (label_flat == 1)).sum())
     den = np.sqrt((tp + fp) * (tp + fn) * (tn + fp) * (tn + fn))
     if den == 0:
         return 0.0
